add settings.py

+++ MeetingSecretary/MS/views.py
@@ -94,7 +94,6 @@ def viewadmingroups(request):
         data_json = item.group.name
         member_results.append(data_json)
     res = {'admin': admin_results, 'member':member_results}
-    print(res)
     res = json.dumps(res)
     mimetype = 'application/json'
     return HttpResponse(res, mimetype)
@@ -105,13 +104,11 @@ def showgroup(request):
     data = Membership.objects.filter(group = group_name)
     admin = Group.objects.get(name = group_name).admin
     results = []
-    print(results)
     for item in data:
         data_json = item.group
         data_json = item.member.username
         results.append(data_json)
     res = {'admin' : admin, 'member' : results}
-    print(res)
     res = json.dumps(res)
     mimetype = 'application/json'
     return HttpResponse(res, mimetype)
@@ -119,7 +116,6 @@ def showgroup(request):
 def deletegroup(request):
     group_name = request.POST.get('groupid')
     operationuser = request.POST.get('operationuser')
-    print(group_name)
     q = Group.objects.get(name = group_name)
     if q.admin == operationuser :
          p  = Membership.objects.filter(group = group_name)
@@ -137,26 +133,36 @@ def deletegroup(request):
 def addnewmember(request):
     group_name = request.POST.get('group_name')
     member_id = request.POST.get('memberid')
+    operationuser = request.POST.get('operationuser')
     group = Group.objects.get(name = group_name)
-    member = User.objects.get(username = member_id)
-    p = Membership(group = group, member = member)
-    p.save()
-    result = 'true'
-    result = json.dumps(result)
+    if group.admin == operationuser :
+        member = User.objects.get(username = member_id)
+        p = Membership(group = group, member = member)
+        p.save()
+        result = 'true'
+    else :
+        result = 'false'
+    res = {'valid': result}
+    res = json.dumps(res)
     mimetype = 'application/json'
-    return HttpResponse(result, mimetype)
+    return HttpResponse(res, mimetype)   
 
 def deletemember(request):
     group_name = request.POST.get('group_name')
     member_id = request.POST.get('memberid')
+    operationuser = request.POST.get('operationuser')
     group = Group.objects.get(name = group_name)
-    member = User.objects.get(username = member_id)
-    p = Membership.objects.get(group = group_name, member = member)
-    p.delete()
-    result = 'true'
-    result = json.dumps(result)
+    if group.admin == operationuser :
+        member = User.objects.get(username = member_id)
+        p = Membership.objects.get(group = group_name, member = member)
+        p.delete()
+        result = 'true'
+    else:
+        result = 'false'
+    res = {'valid': result}
+    res = json.dumps(res)
     mimetype = 'application/json'
-    return HttpResponse(result, mimetype)
+    return HttpResponse(res, mimetype)
 
 #calendar management
 def calendar(request):
